+++ src/components/books/AddBookButton/index.tsx
@@ -1,8 +1,8 @@
 import { Plus } from "lucide-react";
 import Link from "next/link";
-import styles from "@/styles/Buttons.module.css";
+import styles from "./AddBookButton.module.css";
 
-export function AddBook() {
+export function AddBookButton() {
   return (
     <Link href="/create" className={styles.container} scroll={false}>
       <span>책 추가하기</span>
