+++ src/components/ui/Modal/index.tsx
@@ -1,4 +1,4 @@
-import styles from "@/styles/Modal.module.css";
+import styles from "./Modal.module.css";
 
 export default function Modal({ children }: { children: React.ReactNode }) {
   return <div className={styles.container}>{children}</div>;
