+++ src/constants/constants.ts
@@ -1,4 +1,4 @@
-import { InputFieldProps } from "@/components/InputField";
+import { InputFieldProps } from "@/components/ui/InputField";
 
 export const INPUT_FIELDS: InputFieldProps[] = [
   { id: "isbn", label: "ISBN", placeholder: "ISBN을 입력하세요" },
